go-selinux is failing to build on non linux systems
This patch makes go-selinux at least compile on non-linux systems
Basically I stubbed out all of the public content on selinux.go
for non linux systems, and copied the original selinux.go to selinux_linux.go.
